combine style and logic
