--- src/components/utils/KeyboardAwareView/style.js
@@ -1,16 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-const styles = StyleSheet.create({
-  container: {
-    height: '100%',
-    left: 0,
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-    position: 'absolute',
-    top: 0,
-    width: '100%'
-  }
-});
-
-export default styles;
